+++ sdk/search/Azure.Search.Documents/src/Models/SearchSuggestion{T}.cs
@@ -87,8 +87,8 @@ namespace Azure.Search.Documents.Models
             {
                 using Stream stream = element.ToStream();
                 T document = async ?
-                    (T)await serializer.DeserializeAsync(stream, typeof(T)).ConfigureAwait(false) :
-                    (T)serializer.Deserialize(stream, typeof(T));
+                    (T)await serializer.DeserializeAsync(stream, typeof(T), cancellationToken).ConfigureAwait(false) :
+                    (T)serializer.Deserialize(stream, typeof(T), cancellationToken);
                 suggestion.Document = document;
             }
             else